requests: Rename http_requests to requests (#619)
The requests integration is named http-requests because at the time it was created there were some problems with pylint. 

other integrations are using opentelemetry.ext.integration without problems, tests are passing without issue, even renamed.
