+++ docs/examples/opentelemetry-example-app/src/opentelemetry_example_app/flask_example.py
@@ -19,7 +19,7 @@ the requests library to perform downstream requests
 import flask
 import requests
 
-import opentelemetry.ext.http_requests
+import opentelemetry.ext.requests
 from opentelemetry import trace
 from opentelemetry.ext.flask import FlaskInstrumentor
 from opentelemetry.sdk.trace import TracerProvider
@@ -33,7 +33,7 @@ from opentelemetry.sdk.trace.export import (
 # It must be done before instrumenting any library
 trace.set_tracer_provider(TracerProvider())
 
-opentelemetry.ext.http_requests.RequestsInstrumentor().instrument()
+opentelemetry.ext.requests.RequestsInstrumentor().instrument()
 FlaskInstrumentor().instrument()
 
 trace.get_tracer_provider().add_span_processor(
