--- ext/opentelemetry-ext-http-requests/README.rst
@@ -1,23 +1,0 @@
-OpenTelemetry requests Integration
-==================================
-
-|pypi|
-
-.. |pypi| image:: https://badge.fury.io/py/opentelemetry-ext-http-requests.svg
-   :target: https://pypi.org/project/opentelemetry-ext-http-requests/
-
-This library allows tracing HTTP requests made by the
-`requests <https://requests.readthedocs.io/en/master/>`_ library.
-
-Installation
-------------
-
-::
-
-     pip install opentelemetry-ext-http-requests
-
-References
-----------
-
-* `OpenTelemetry requests Integration <https://opentelemetry-python.readthedocs.io/en/latest/ext/http_requests/http_requests..html>`_
-* `OpenTelemetry Project <https://opentelemetry.io/>`_
