+++ ext/opentelemetry-ext-requests/MANIFEST.in
