+++ ext/opentelemetry-ext-requests/README.rst
@@ -1,0 +1,23 @@
+OpenTelemetry requests Integration
+==================================
+
+|pypi|
+
+.. |pypi| image:: https://badge.fury.io/py/opentelemetry-ext-requests.svg
+   :target: https://pypi.org/project/opentelemetry-ext-requests/
+
+This library allows tracing HTTP requests made by the
+`requests <https://requests.readthedocs.io/en/master/>`_ library.
+
+Installation
+------------
+
+::
+
+     pip install opentelemetry-ext-requests
+
+References
+----------
+
+* `OpenTelemetry requests Integration <https://opentelemetry-python.readthedocs.io/en/latest/ext/requests/requests.html>`_
+* `OpenTelemetry Project <https://opentelemetry.io/>`_
