+++ ext/opentelemetry-ext-requests/setup.cfg
@@ -13,13 +13,13 @@
 # limitations under the License.
 #
 [metadata]
-name = opentelemetry-ext-http-requests
+name = opentelemetry-ext-requests
 description = OpenTelemetry requests integration
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 author = OpenTelemetry Authors
 author_email = cncf-opentelemetry-contributors@lists.cncf.io
-url = https://github.com/open-telemetry/opentelemetry-python/tree/master/ext/opentelemetry-ext-http-requests
+url = https://github.com/open-telemetry/opentelemetry-python/tree/master/ext/opentelemetry-ext-requests
 platforms = any
 license = Apache-2.0
 classifiers =
@@ -54,4 +54,4 @@ where = src
 
 [options.entry_points]
 opentelemetry_instrumentor =
-    requests = opentelemetry.ext.http_requests:RequestsInstrumentor
+    requests = opentelemetry.ext.requests:RequestsInstrumentor
