+++ ext/opentelemetry-ext-requests/src/opentelemetry/ext/requests/version.py
