+++ tox.ini
@@ -36,9 +36,9 @@ envlist =
     py3{4,5,6,7,8}-test-ext-flask
     pypy3-test-ext-flask
 
-    ; opentelemetry-ext-http-requests
-    py3{4,5,6,7,8}-test-ext-http-requests
-    pypy3-test-ext-http-requests
+    ; opentelemetry-ext-requests
+    py3{4,5,6,7,8}-test-ext-requests
+    pypy3-test-ext-requests
 
     ; opentelemetry-ext-jaeger
     py3{4,5,6,7,8}-test-ext-jaeger
@@ -124,7 +124,7 @@ changedir =
   test-sdk: opentelemetry-sdk/tests
   test-auto-instrumentation: opentelemetry-auto-instrumentation/tests
   test-ext-grpc: ext/opentelemetry-ext-grpc/tests
-  test-ext-http-requests: ext/opentelemetry-ext-http-requests/tests
+  test-ext-requests: ext/opentelemetry-ext-requests/tests
   test-ext-jaeger: ext/opentelemetry-ext-jaeger/tests
   test-ext-dbapi: ext/opentelemetry-ext-dbapi/tests
   test-ext-mysql: ext/opentelemetry-ext-mysql/tests
@@ -156,13 +156,13 @@ commands_pre =
   test-auto-instrumentation: pip install {toxinidir}/opentelemetry-auto-instrumentation
 
   example-app: pip install {toxinidir}/opentelemetry-auto-instrumentation
-  example-app: pip install {toxinidir}/ext/opentelemetry-ext-http-requests
+  example-app: pip install {toxinidir}/ext/opentelemetry-ext-requests
   example-app: pip install {toxinidir}/ext/opentelemetry-ext-wsgi
   example-app: pip install {toxinidir}/ext/opentelemetry-ext-flask
   example-app: pip install {toxinidir}/docs/examples/opentelemetry-example-app
 
   example-http: pip install -e {toxinidir}/opentelemetry-auto-instrumentation
-  example-http: pip install -e {toxinidir}/ext/opentelemetry-ext-http-requests
+  example-http: pip install -e {toxinidir}/ext/opentelemetry-ext-requests
   example-http: pip install -e {toxinidir}/ext/opentelemetry-ext-wsgi
   example-http: pip install -r {toxinidir}/docs/examples/http/requirements.txt
 
@@ -195,8 +195,8 @@ commands_pre =
   redis: pip install {toxinidir}/opentelemetry-auto-instrumentation
   redis: pip install {toxinidir}/ext/opentelemetry-ext-redis[test]
 
-  http-requests: pip install {toxinidir}/opentelemetry-auto-instrumentation
-  http-requests: pip install {toxinidir}/ext/opentelemetry-ext-http-requests[test]
+  requests: pip install {toxinidir}/opentelemetry-auto-instrumentation
+  requests: pip install {toxinidir}/ext/opentelemetry-ext-requests[test]
 
   jaeger: pip install {toxinidir}/ext/opentelemetry-ext-jaeger
 
@@ -271,7 +271,7 @@ commands_pre =
               -e {toxinidir}/opentelemetry-auto-instrumentation \
               -e {toxinidir}/opentelemetry-sdk \
               -e {toxinidir}/opentelemetry-auto-instrumentation \
-              -e {toxinidir}/ext/opentelemetry-ext-http-requests \
+              -e {toxinidir}/ext/opentelemetry-ext-requests \
               -e {toxinidir}/ext/opentelemetry-ext-wsgi \
               -e {toxinidir}/ext/opentelemetry-ext-flask
